Updated randomizer logic
The randomizer now takes in consideration of the new filter 'Include Side Characters'. If the toggle is on and side characters are selected, those characters will be added to the randomizer pool. If the toggle is off, the characters remain 'Checked' but hidden and also excluded from the randomizer pool of characters.